+++ pkg/deployment/resources/internal_exporter.go
@@ -1,5 +1,5 @@
 //
-// Copyright 2016-2022 ArangoDB GmbH, Cologne, Germany
+// Copyright 2016-2023 ArangoDB GmbH, Cologne, Germany
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -55,8 +55,8 @@ func ArangodbInternalExporterContainer(image string, args []string, livenessProb
 			},
 		},
 		Resources:       k8sutil.ExtractPodResourceRequirement(resources),
+		SecurityContext: k8sutil.CreateSecurityContext(groupSpec.SecurityContext),
 		ImagePullPolicy: core.PullIfNotPresent,
-		SecurityContext: groupSpec.SecurityContext.NewSecurityContext(),
 		VolumeMounts:    []core.VolumeMount{k8sutil.LifecycleVolumeMount()},
 	}